+++ bin/app/src/common/i18n.js
@@ -1,0 +1,9 @@
+
+/**
+ * Returns the translated string to the currently active language.
+ * @param {string} value
+ * @returns {string}
+ */
+export function tr (value) {
+    return value;
+}