+++ bin/app/src/common/messages.js
@@ -1,0 +1,12 @@
+
+self.messages =
+{
+    'required:true': 'Field is required',
+    'pattern:identifier': 'Value must be an identifier',
+    'pattern:integer': 'Value must be an integer',
+    'pattern:email': 'Invalid email address',
+    'min-length:5': 'Field must be at least 5 characters long',
+    'min-length:4': 'Field must be at least 4 characters long',
+    'max-length:8': 'Field must be at most 8 characters long',
+    'max-length:50': 'Field must be at most 50 characters long',
+};